+++ packages/xchain-bitcoin/src/clientKeystore.ts
@@ -1,0 +1,112 @@
+import { FeeOption, FeeRate, TxHash, TxParams, checkFeeBounds } from '@xchainjs/xchain-client'
+import { getSeed } from '@xchainjs/xchain-crypto'
+import { Address } from '@xchainjs/xchain-util'
+import * as Bitcoin from 'bitcoinjs-lib'
+
+import { Client } from './client'
+import * as Utils from './utils'
+
+/**
+ * Custom Bitcoin client
+ */
+class ClientKeystore extends Client {
+  /**
+   * @deprecated this function eventually will be removed use getAddressAsync instead
+   */
+  getAddress(index = 0): Address {
+    if (index < 0) {
+      throw new Error('index must be greater than zero')
+    }
+    if (this.phrase) {
+      const btcNetwork = Utils.btcNetwork(this.network)
+      const btcKeys = this.getBtcKeys(this.phrase, index)
+
+      const { address } = Bitcoin.payments.p2wpkh({
+        pubkey: btcKeys.publicKey,
+        network: btcNetwork,
+      })
+      if (!address) {
+        throw new Error('Address not defined')
+      }
+      return address
+    }
+    throw new Error('Phrase must be provided')
+  }
+
+  /**
+   * Get the current address.
+   *
+   * Generates a network-specific key-pair by first converting the buffer to a Wallet-Import-Format (WIF)
+   * The address is then decoded into type P2WPKH and returned.
+   *
+   * @returns {Address} The current address.
+   *
+   * @throws {"Phrase must be provided"} Thrown if phrase has not been set before.
+   * @throws {"Address not defined"} Thrown if failed creating account from phrase.
+   */
+  async getAddressAsync(index = 0): Promise<string> {
+    return this.getAddress(index)
+  }
+
+  /**
+   * @private
+   * Get private key.
+   *
+   * Private function to get keyPair from the this.phrase
+   *
+   * @param {string} phrase The phrase to be used for generating privkey
+   * @returns {ECPairInterface} The privkey generated from the given phrase
+   *
+   * @throws {"Could not get private key from phrase"} Throws an error if failed creating BTC keys from the given phrase
+   * */
+  private getBtcKeys(phrase: string, index = 0): Bitcoin.ECPair.ECPairInterface {
+    const btcNetwork = Utils.btcNetwork(this.network)
+
+    const seed = getSeed(phrase)
+    const master = Bitcoin.bip32.fromSeed(seed, btcNetwork).derivePath(this.getFullDerivationPath(index))
+
+    if (!master.privateKey) {
+      throw new Error('Could not get private key from phrase')
+    }
+
+    return Bitcoin.ECPair.fromPrivateKey(master.privateKey, { network: btcNetwork })
+  }
+
+  /**
+   * Transfer BTC.
+   *
+   * @param {TxParams&FeeRate} params The transfer options.
+   * @returns {TxHash} The transaction hash.
+   *
+   * @throws {"memo too long"} Thrown if memo longer than  80 chars.
+   */
+  async transfer(params: TxParams & { feeRate?: FeeRate }): Promise<TxHash> {
+    // set the default fee rate to `fast`
+    const feeRate = params.feeRate || (await this.getFeeRates())[FeeOption.Fast]
+    checkFeeBounds(this.feeBounds, feeRate)
+
+    const fromAddressIndex = params?.walletIndex || 0
+    const { rawUnsignedTx } = await this.prepareTx({ ...params, sender: this.getAddress(fromAddressIndex), feeRate })
+
+    const btcKeys = this.getBtcKeys(this.phrase, fromAddressIndex)
+    const psbt = Bitcoin.Psbt.fromBase64(rawUnsignedTx)
+
+    psbt.signAllInputs(btcKeys) // Sign all inputs
+    psbt.finalizeAllInputs() // Finalise inputs
+
+    const txHex = psbt.extractTransaction().toHex() // TX extracted and formatted to hex
+
+    const txHash = psbt.extractTransaction().getId()
+    try {
+      const txId = await this.roundRobinBroadcastTx(txHex)
+      return txId
+    } catch (err) {
+      const error = `Server error, please check explorer for tx confirmation ${this.explorerProviders[
+        this.network
+      ].getExplorerTxUrl(txHash)}`
+      return error
+    }
+  }
+}
+
+export { ClientKeystore }