+++ packages/xchain-client/src/types.ts
@@ -1,5 +1,7 @@
 import { Address, Asset, BaseAmount } from '@xchainjs/xchain-util'
 
+import { UTXO } from './provider-types'
+
 export enum Network {
   Mainnet = 'mainnet',
   Stagenet = 'stagenet',
@@ -102,6 +104,7 @@ export type XChainClientParams = {
 
 export type PreparedTx = {
   rawUnsignedTx: string
+  utxos?: UTXO[]
 }
 
 export interface XChainClient {